Address non-test feedback.
