+++ src/EditorFeatures/Core/Implementation/CommentSelection/CommentTrackingSpan.cs
@@ -7,32 +7,35 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.CommentSelection
     /// <summary>
     /// Wrapper around a TextSpan that holds extra data used to create a tracking span.
     /// </summary>
-    internal struct CommentTrackingSpan
+    internal readonly struct CommentTrackingSpan
     {
         public TextSpan TrackingTextSpan { get; }
 
         // In some cases, the tracking span needs to be adjusted by a specific amount after the changes have been applied.
         // These fields store the amount to adjust the span by after edits have been applied.
-        public int AmountToAddToStart { get; }
-        public int AmountToAddToEnd { get; }
+        // e.g. The selection begins in a comment -
+        //    /*Com[|ment*/ int i = 1;|] -> /*Com*//*ment*/ int i = 1;*/
+        // There are two new comment markers added inside the original comment, only the second should be selected.
+        public int AmountToAddToTrackingSpanStart { get; }
+        public int AmountToAddToTrackingSpanEnd { get; }
 
         public CommentTrackingSpan(TextSpan trackingTextSpan)
         {
             TrackingTextSpan = trackingTextSpan;
-            AmountToAddToStart = 0;
-            AmountToAddToEnd = 0;
+            AmountToAddToTrackingSpanStart = 0;
+            AmountToAddToTrackingSpanEnd = 0;
         }
 
         public CommentTrackingSpan(TextSpan trackingTextSpan, int amountToAddToStart, int amountToAddToEnd)
         {
             TrackingTextSpan = trackingTextSpan;
-            AmountToAddToStart = amountToAddToStart;
-            AmountToAddToEnd = amountToAddToEnd;
+            AmountToAddToTrackingSpanStart = amountToAddToStart;
+            AmountToAddToTrackingSpanEnd = amountToAddToEnd;
         }
 
         public bool HasPostApplyChanges()
         {
-            return AmountToAddToStart != 0 || AmountToAddToEnd != 0;
+            return AmountToAddToTrackingSpanStart != 0 || AmountToAddToTrackingSpanEnd != 0;
         }
     }
 }
