+++ src/EditorFeatures/Core/Implementation/CommentSelection/IToggleBlockCommentDocumentDataProvider.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using Microsoft.CodeAnalysis.Text;
 
 namespace Microsoft.CodeAnalysis.Editor.Implementation.CommentSelection
@@ -15,6 +16,6 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.CommentSelection
         /// <summary>
         /// Gets all block comments in a particular document.
         /// </summary>
-        IEnumerable<TextSpan> GetBlockCommentsInDocument();
+        ImmutableArray<TextSpan> GetBlockCommentsInDocument();
     }
 }
